+++ CHANGELOG.md
@@ -1,10 +1,15 @@
 # opencred-platform Changelog
 
+## 10.0.4 - 2026-03-XX
+
+### Fixed
+- Fix config and frontend mismatch around redirectUri context item. OIDC login
+  success shows an i18n manual-continue line after a short delay.
+
 ## 10.0.3 - 2026-03-18
 
 ### Fixed
 - Fixes default OID4VP protocol expression to `opencred.options.OID4VPdefault`.
-
 
 ## 10.0.2 - 2026-03-11